% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BVS_HHS_MVN.R
\name{fit_BVS_HHS_MVN}
\alias{fit_BVS_HHS_MVN}
\title{Update parameters for BVS-HHS-MVN model}
\usage{
fit_BVS_HHS_MVN(
  niter = 20,
  burn_in = 2,
  thin = 1,
  n,
  K,
  Y,
  W,
  n_all_par,
  J,
  M,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = matrix(0.5, nrow = J, ncol = K),
  tausq_beta_init = rep(1, J),
  lambdasq_gamma_init = matrix(0.5, nrow = M, ncol = K),
  tausq_gamma_init = rep(1, M),
  lambdasq_delta_init = matrix(0.5, nrow = J * M, ncol = K),
  tausq_delta_init = rep(1, J * M),
  psi_beta_init = matrix(0.5, nrow = J, ncol = K),
  psi_gamma_init = matrix(0.5, nrow = M, ncol = K),
  psi_delta_init = matrix(0.5, nrow = J * M, ncol = K),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init = diag(K),
  nu_0,
  Psi_0
)
}
\description{
Update parameters for BVS-HHS-MVN model
}
