% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_BVS_HRHS_SI_MVN_cov_modify.R
\name{fit_BVS_HRHS_SI_MVN_cov_modify}
\alias{fit_BVS_HRHS_SI_MVN_cov_modify}
\title{Update parameters for BVS-HRHS-SI-MVN_cov model}
\usage{
fit_BVS_HRHS_SI_MVN_cov_modify(
  niter = 20,
  burn_in = 2,
  thin = 1,
  n,
  K,
  Y,
  W,
  WTW,
  n_all_par,
  J,
  M,
  O,
  c = 2.5,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = rep(0.5, J),
  tausq_beta_init = 1,
  lambdasq_gamma_init = rep(0.5, M),
  tausq_gamma_init = 1,
  lambdasq_delta_init = rep(0.5, J * M),
  tausq_delta_init = 1,
  psi_beta_init = rep(0.5, J),
  psi_gamma_init = rep(0.5, M),
  psi_delta_init = rep(0.5, J * M),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init = diag(K),
  nu_0,
  Psi_0,
  omegasq_beta_lambdasq = 0.25,
  omegasq_beta_tausq = 0.25,
  omegasq_gamma_lambdasq = 0.25,
  omegasq_gamma_tausq = 0.25,
  omegasq_delta_lambdasq = 0.25,
  omegasq_delta_tausq = 0.25,
  accept_lambdasq_beta_init = rep(1, times = J),
  accept_tausq_beta_init = 1,
  accept_lambdasq_gamma_init = rep(1, times = M),
  accept_tausq_gamma_init = 1,
  accept_lambdasq_delta_init = rep(1, times = J * M),
  accept_tausq_delta_init = 1,
  sigmasq_varphi = 10
)
}
\description{
Update parameters for BVS-HRHS-SI-MVN_cov model
}
