% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSHHS_SI_MVN.R
\name{fit_BVSHHS_SI_MVN}
\alias{fit_BVSHHS_SI_MVN}
\title{Update parameters for BVSHHS-SI-MVN model}
\usage{
fit_BVSHHS_SI_MVN(
  niter = 6000,
  burn_in = 1000,
  thin = 5,
  n,
  K,
  Y,
  W,
  n_all_par,
  J,
  M,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = rep(0.5, J),
  tausq_beta_init = 1,
  lambdasq_gamma_init = rep(0.5, M),
  tausq_gamma_init = 1,
  lambdasq_delta_init = rep(0.5, J * M),
  tausq_delta_init = 1,
  psi_beta_init = rep(0.5, J),
  psi_gamma_init = rep(0.5, M),
  psi_delta_init = rep(0.5, J * M),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init,
  nu_0,
  Psi_0
)
}
\description{
Update parameters for BVSHHS-SI-MVN model
}
