% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVSHRHS_SI_MVN.R
\name{fit_BVSHRHS_SI_MVN}
\alias{fit_BVSHRHS_SI_MVN}
\title{Update parameters for BVSHRHS-SI-MVN model}
\usage{
fit_BVSHRHS_SI_MVN(
  niter = 20,
  burn_in = 2,
  thin = 1,
  n,
  K,
  Y,
  W,
  n_all_par,
  J,
  M,
  c = 2.5,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = rep(0.5, J),
  tausq_beta_init = 1,
  lambdasq_gamma_init = rep(0.5, M),
  tausq_gamma_init = 1,
  lambdasq_delta_init = rep(0.5, J * M),
  tausq_delta_init = 1,
  psi_beta_init = rep(0.5, J),
  psi_gamma_init = rep(0.5, M),
  psi_delta_init = rep(0.5, J * M),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init,
  nu_0,
  Psi_0,
  omegasq_beta_lambdasq = 0.25,
  omegasq_beta_tausq = 0.25,
  omegasq_gamma_lambdasq = 0.25,
  omegasq_gamma_tausq = 0.25,
  omegasq_delta_lambdasq = 0.25,
  omegasq_delta_tausq = 0.25,
  accept_lambdasq_beta_init = rep(1, times = J),
  accept_tausq_beta_init = 1,
  accept_lambdasq_gamma_init = rep(1, times = M),
  accept_tausq_gamma_init = 1,
  accept_lambdasq_delta_init = rep(1, times = J * M),
  accept_tausq_delta_init = 1
)
}
\description{
Update parameters for BVSHRHS-SI-MVN model
}
