% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_HHS_MVN_cov_modify3.R
\name{fit_HHS_MVN_cov_modify3}
\alias{fit_HHS_MVN_cov_modify3}
\title{Update parameters for HHS-MVN-cov model}
\usage{
fit_HHS_MVN_cov_modify3(
  niter = 6000,
  burn_in = 1000,
  thin = 5,
  n,
  K,
  Y,
  W,
  WTW,
  n_all_par,
  J,
  M,
  O,
  sigmasq_alpha = 100,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = matrix(0.5, nrow = J, ncol = K),
  tausq_beta_init = rep(1, J),
  lambdasq_gamma_init = matrix(0.5, nrow = M, ncol = K),
  tausq_gamma_init = rep(1, M),
  lambdasq_delta_init = matrix(0.5, nrow = J * M, ncol = K),
  tausq_delta_init = rep(1, J * M),
  psi_beta_init = matrix(0.5, nrow = J, ncol = K),
  psi_gamma_init = matrix(0.5, nrow = M, ncol = K),
  psi_delta_init = matrix(0.5, nrow = J * M, ncol = K),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init = diag(K),
  nu_0,
  Psi_0,
  sigmasq_varphi = 10
)
}
\description{
Update parameters for HHS-MVN-cov model
}
