% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NHHSSI_MVN_cov_modify3.R
\name{fit_NHHSSI_MVN_cov_modify3}
\alias{fit_NHHSSI_MVN_cov_modify3}
\title{Update parameters for NHHSSI-MVN-cov model}
\usage{
fit_NHHSSI_MVN_cov_modify3(
  niter = 20,
  burn_in = 2,
  thin = 1,
  n,
  K,
  Y,
  W,
  WTW,
  n_all_par,
  J,
  M,
  O,
  sigmasq_alpha = 100,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  lambdasq_beta_init = rep(0.5, J),
  tausq_beta_init = 1,
  lambdasq_gamma_init = rep(0.5, M),
  tausq_gamma_init = 1,
  lambdasq_delta_init = rep(0.5, J * M),
  tausq_delta_init = 1,
  psi_beta_init = rep(0.5, J),
  psi_gamma_init = rep(0.5, M),
  psi_delta_init = rep(0.5, J * M),
  xi_beta_init = 1,
  xi_gamma_init = 1,
  xi_delta_init = 1,
  Sigma_init,
  nu_0,
  Psi_0,
  sigmasq_varphi = 10
)
}
\description{
Update parameters for NHHSSI-MVN-cov model
}
