% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NP_MVN_cov.R
\name{fit_NP_MVN_cov}
\alias{fit_NP_MVN_cov}
\title{Update parameters for NP-MVN-cov model}
\usage{
fit_NP_MVN_cov(
  niter = 20,
  burn_in = 2,
  thin = 1,
  n,
  K,
  Y,
  W,
  n_all_par,
  J,
  M,
  O,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  Sigma_init = matrix(0.5, nrow = K, ncol = K),
  nu_0,
  Psi_0,
  sigmasq_varphi = 10
)
}
\description{
Update parameters for NP-MVN-cov model
}
