% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_NP_MVN_cov_modify3.R
\name{fit_NP_MVN_cov_modify3}
\alias{fit_NP_MVN_cov_modify3}
\title{Update parameters for NP-MVN-cov model}
\usage{
fit_NP_MVN_cov_modify3(
  niter = 6000,
  burn_in = 1000,
  thin = 5,
  n,
  K,
  Y,
  W,
  WTW,
  n_all_par,
  J,
  M,
  O,
  sigmasq_alpha = 100,
  theta_init = matrix(0.5, nrow = n_all_par, ncol = K),
  Sigma_init = diag(K),
  nu_0,
  Psi_0,
  sigmasq_varphi = 10
)
}
\description{
Update parameters for NP-MVN-cov model
}
