% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVS_MVN_data_simulation.R
\name{Sim_data_BVS_real_elevated_active_coef}
\alias{Sim_data_BVS_real_elevated_active_coef}
\title{Smulate data}
\usage{
Sim_data_BVS_real_elevated_active_coef(
  K = 5,
  n = 1000,
  J = 44,
  M = 17,
  O = 6,
  x,
  z,
  d,
  Sigma
)
}
\description{
Smulate data
}
