% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVS_MVN_data_simulation.R
\name{Sim_data_BVS_real_sim_var}
\alias{Sim_data_BVS_real_sim_var}
\title{all variables are simulated, mimicking real data by number of variable
Sigma=error variance is same as the real data}
\usage{
Sim_data_BVS_real_sim_var(K = 4, n = 1301, J = 44, M = 17, O = 6, Sigma)
}
\description{
all variables are simulated, mimicking real data by number of variable
Sigma=error variance is same as the real data
}
